Restructured inspect related samples - 2
